⚡ Bolt: Replace f32::powf with f32::powi for integer exponents (#1669)
**What:**
Replaced all occurrences of `x.powf(n.0)` with `x.powi(n)` across the engine where `n` is a known whole number (e.g. `powf(10.0)` -> `powi(10)`).

**Why:**
Using `powf` with an integer is significantly slower than `powi` due to the underlying C-math fallback for fractional/negative powers.

**Impact:**
Measurably faster math evaluations, particularly in specular shading loops and easing functions.

**Measurement:**
Micro-benchmarks show an improvement when executing tight loop `powi` versus `powf`. I also resolved a leftover unused `fast_sin_cos` reference that blocked compilation, and fixed numerous broken example usages resulting from previous WindowApp API changes so that `cargo check --examples --all-features` and `cargo test` finally succeed cleanly.

+++ examples/gamma_demo.rs
@@ -57,7 +57,7 @@ fn main() {
     // Draw some simple shapes/gradient
     for y in 0..height {
         for x in 0..width {
-            let val = ((x + y) % 256) as u32;
+            let val = (x + y) % 256 ;
             let color = 0xFF00_0000 | (val << 16) | (val << 8) | val;
             fb.set_pixel(x as i32, y as i32, color);
         }
